Job features negative tests

+++ app/Http/Controllers/JobController.php
@@ -53,11 +53,12 @@ class JobController extends Controller
         $this->authorize('create', Job::class);
 
         $validatedJob = $request->validate([
-            'title' => 'required|string|max:255',
-            'description' => 'required|string|max:5000',
+            'title' => 'required|string|min:5|max:255',
+            'description' => 'required|string|min:5|max:5000',
             'min_price' => 'required|numeric|min:0',
             'max_price' => 'required|numeric|min:0|gt:min_price',
-            'type' => 'required|in:hourly,fixed_price'
+            'type' => 'required|in:hourly,fixed_price',
+            'skills' => 'required|array|exists:skills,id',
         ]);
 
         $skills = $request->input('skills');
@@ -117,6 +118,7 @@ class JobController extends Controller
             'min_price' => 'numeric|min:0',
             'max_price' => 'numeric|min:0|gt:min_price',
             'type' => 'in:hourly,fixed_price',
+            'skills' => 'array|exists:skills,id',
         ]);
 
         $skills = $request->input('skills');
@@ -134,6 +136,8 @@ class JobController extends Controller
      */
     public function destroy(Job $job)
     {
+        $this->authorize('delete', $job);
+
         $job->delete();
         return redirect(route('jobs.index'), 303);
     }
